API for Vehicle Makes. Added angular application. Implemented GET for vehicle makes.

+++ VehicleApp.Service/Repository/VehicleMakeRepository.cs
@@ -66,9 +66,13 @@ namespace VehicleApp.Repository
         }
         public void DeleteConfirmed(int? Id)
         {
+
             VehicleMake vehicleMake = context.Makes.Find(Id);
-            context.Makes.Remove(vehicleMake);
-            context.SaveChanges();
+            if (vehicleMake != null)
+            {
+                context.Makes.Remove(vehicleMake);
+                context.SaveChanges();
+            }
         }
     }
 }
